refactor(ooda_actions/advance_goal): split into mod+spawn+subordinate (#1266)
advance_goal.rs 657 -> mod.rs 105 + spawn.rs 303 + subordinate.rs 285.
spawn.rs: dispatch_spawn_engineer, find_live_engineer_for_goal, helpers.
subordinate.rs: advance_goal_with_subordinate, cleanup_engineer_worktree_for_goal,
validate_subordinate_completion.
